+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-* 21.01
+* 21.0.0
   - Plotter2D now takes 4D datasets and slice infomation as input
   - TIGRE reconstruction package wrapped for cone-beam tomography
+  - Datacontainers have get_slice method which returns a dataset with a single slice of the data
+  - Datacontainers have reorder method which returns a dataset with the axes reordered
 
 * 20.11.2
   - fixed windows build